requirements.txt
dspy>=3.1.3
google-genai
python-dotenv
